+++ dffml/db/base.py
@@ -4,8 +4,8 @@ import functools
 import collections
 from typing import Any, List, Optional, Dict, Tuple, Union, AsyncIterator
 
-from dffml.df.base import BaseDataFlowObject, BaseDataFlowObjectContext
-from dffml.util.entrypoint import base_entry_point
+from ..df.base import BaseDataFlowObject, BaseDataFlowObjectContext
+from ..util.entrypoint import base_entry_point
 
 
 Condition = collections.namedtuple(