+++ dffml/operation/preprocess.py
@@ -1,6 +1,7 @@
 import ast
-from dffml.df.types import Definition
-from dffml.df.base import op
+
+from ..df.types import Definition
+from ..df.base import op
 
 
 # Definitions
@@ -32,6 +33,9 @@ async def literal_eval(str_to_eval: str):
 
     The following example shows how to use literal_eval.
 
+    >>> import asyncio
+    >>> from dffml import *
+    >>>
     >>> dataflow = DataFlow.auto(literal_eval, GetSingle)
     >>> dataflow.seed.append(
     ...    Input(