+++ examples/shouldi/shouldi/golangci_lint.py
@@ -2,8 +2,7 @@ import json
 import asyncio
 from typing import Dict, Any
 
-from dffml.df.base import op
-from dffml.df.types import Definition
+from dffml import op, Definition
 
 package_src_dir = Definition(name="package_src_dir", primitive="str")
 golangci_lint_output = Definition(