+++ examples/shouldi/shouldi/npm_audit.py
@@ -2,8 +2,8 @@ import json
 import asyncio
 from typing import Dict, Any
 
-from dffml.df.base import op
-from dffml.df.types import Definition
+from dffml import op
+from dffml import Definition
 
 package_src_dir = Definition(name="package_src_dir", primitive="str")
 npm_audit_output = Definition(
@@ -31,8 +31,8 @@ async def run_npm_audit(pkg: str) -> Dict[str, Any]:
         stderr=asyncio.subprocess.PIPE,
     )
 
-    _, stderr = await proc.communicate()
-    if proc.returncode != 0:
+    stdout, stderr = await proc.communicate()
+    if proc.returncode != 0 and stderr:
         raise NPMAuditError(stderr.decode())
 
     npm_audit_op = stdout.decode()