+++ tests/test_noasync.py
@@ -2,15 +2,12 @@
 This file contains integration tests for the high level (very abstract) APIs.
 """
 import importlib
-import contextlib
 
 from dffml.record import Record
-from dffml import run, train, accuracy, predict, save, load
+from dffml.noasync import train, accuracy, predict
 from dffml.source.csv import CSVSource
 from dffml.feature.feature import Features, DefFeature
 from dffml.util.asynctestcase import IntegrationCLITestCase
-
-from .test_df import TestOrchestrator, DATAFLOW
 
 FEATURE_NAMES = ["Years", "Expertise", "Trust", "Salary"]
 
@@ -26,8 +23,6 @@ class TestML(IntegrationCLITestCase):
 
     async def setUp(self):
         await super().setUp()
-        save_and_load_file = self.mktempfile() + ".csv"
-        setattr(self, "save_and_load", save_and_load_file)
         self.train_data = [
             [0, 1, 0.2, 10],
             [1, 3, 0.4, 20],
@@ -48,62 +43,7 @@ class TestML(IntegrationCLITestCase):
             setattr(self, f"{use}_filename", filename)
             await self.populate_source(CSVSource, *records, filename=filename)
 
-    async def test_save_and_load(self):
-        source = CSVSource(
-            filename=self.save_and_load, allowempty=True, readwrite=True
-        )
-        await save(
-            source,
-            Record(
-                "1",
-                data={
-                    "features": {"A": 0, "B": 1},
-                    "prediction": {"C": {"value": 1, "confidence": 1.0}},
-                },
-            ),
-            Record(
-                "2",
-                data={
-                    "features": {"A": 3, "B": 4},
-                    "prediction": {"C": {"value": 2, "confidence": 1.0}},
-                },
-            ),
-        )
-        # All records in source
-        results = [record.export() async for record in load(source)]
-        self.assertEqual(
-            results,
-            [
-                {
-                    "key": "1",
-                    "features": {"A": 0, "B": 1},
-                    "prediction": {"C": {"confidence": 1.0, "value": "1"}},
-                    "extra": {},
-                },
-                {
-                    "key": "2",
-                    "features": {"A": 3, "B": 4},
-                    "prediction": {"C": {"confidence": 1.0, "value": "2"}},
-                    "extra": {},
-                },
-            ],
-        )
-
-        # For specific records in a source
-        results = [record.export() async for record in load(source, "1")]
-        self.assertEqual(
-            results,
-            [
-                {
-                    "key": "1",
-                    "features": {"A": 0, "B": 1},
-                    "prediction": {"C": {"confidence": 1.0, "value": "1"}},
-                    "extra": {},
-                }
-            ],
-        )
-
-    async def test_predict(self):
+    def test_predict(self):
         self.required_plugins("dffml-model-scikit")
         # Import SciKit modules
         dffml_model_scikit = importlib.import_module("dffml_model_scikit")
@@ -123,21 +63,12 @@ class TestML(IntegrationCLITestCase):
         predict_data = CSVSource(filename=self.predict_filename)
 
         # Train the model
-        await train(model, training_data)
+        train(model, training_data)
         # Assess accuracy
-        await accuracy(model, test_data)
+        accuracy(model, test_data)
         # Make prediction
         predictions = [
-            prediction async for prediction in predict(model, predict_data)
+            prediction for prediction in predict(model, predict_data)
         ]
         self.assertEqual(predictions[0][2]["Salary"]["value"], 70)
         self.assertEqual(predictions[1][2]["Salary"]["value"], 80)
-
-
-class TestDataFlow(TestOrchestrator):
-    @contextlib.asynccontextmanager
-    async def create_octx(self):
-        yield None
-
-    def run_dataflow(self, _octx, *inputs):
-        return run(DATAFLOW, *inputs)